move gif creation logic into @noita-explorer/tools
