+++ packages/tools/package.json
@@ -34,6 +34,7 @@
     "@jest/globals": "^29.7.0",
     "@types/color": "^4.2.0",
     "@types/luaparse": "^0.2.12",
+    "@types/omggif": "^1.0.5",
     "@types/xml2js": "^0.4.14",
     "barrelsby": "^2.8.1",
     "jest": "^29.7.0",
@@ -45,10 +46,12 @@
   "dependencies": {
     "@noita-explorer/model": "*",
     "buffer": "^6.0.3",
+    "canvas": "^3.1.2",
     "color": "^4.2.3",
     "events": "^3.3.0",
+    "jimp": "^1.6.0",
     "luaparse": "^0.3.1",
-    "xml2js": "^0.6.2",
-    "jimp": "^1.6.0"
+    "omggif": "^1.0.10",
+    "xml2js": "^0.6.2"
   }
 }
