+++ packages/tools/src/common/gif/gif.browser.ts
@@ -1,0 +1,7 @@
+import { CreateGifReturnType, GifHelpersType } from './gif.types.ts';
+
+async function createGif(): Promise<CreateGifReturnType> {
+  throw new Error('createGif is not implemented for browser environment');
+}
+
+export const gifHelper: GifHelpersType = { createGif };
